fix(expression): 修复表达式小数位丢失问题

+++ luck-report-server/luck-report-core/src/main/java/com/luck/report/core/export/excel/high/builder/ExcelBuilderWithPaging.java
@@ -150,7 +150,7 @@ public class ExcelBuilderWithPaging extends ExcelBuilder {
                                 cell.setCellValue((String) obj);
                             } else if (obj instanceof Number) {
                                 BigDecimal bigDecimal = Utils.toBigDecimal(obj);
-                                cell.setCellValue(bigDecimal.floatValue());
+                                cell.setCellValue(bigDecimal.doubleValue());
                             } else if (obj instanceof Boolean) {
                                 cell.setCellValue((Boolean) obj);
                             } else if (obj instanceof Image) {
