+++ luck-report-server/luck-report-core/src/main/java/com/luck/report/core/export/excel/low/Excel97Producer.java
@@ -155,7 +155,7 @@ public class Excel97Producer {
                                     cell.setCellValue((String) obj);
                                 } else if (obj instanceof Number) {
                                     BigDecimal bigDecimal = Utils.toBigDecimal(obj);
-                                    cell.setCellValue(bigDecimal.floatValue());
+                                    cell.setCellValue(bigDecimal.doubleValue());
                                 } else if (obj instanceof Boolean) {
                                     cell.setCellValue((Boolean) obj);
                                 } else if (obj instanceof Image) {
@@ -304,7 +304,7 @@ public class Excel97Producer {
                                 cell.setCellValue((String) obj);
                             } else if (obj instanceof Number) {
                                 BigDecimal bigDecimal = Utils.toBigDecimal(obj);
-                                cell.setCellValue(bigDecimal.floatValue());
+                                cell.setCellValue(bigDecimal.doubleValue());
                             } else if (obj instanceof Boolean) {
                                 cell.setCellValue((Boolean) obj);
                             } else if (obj instanceof Image) {
